refactor: reorganize component imports
